requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
fpdf==1.7.2
